Extract TimedChallengeInstructions
