+++ mon-pix/app/templates/assessments/challenge.hbs
@@ -18,18 +18,25 @@
 
   <main class="assessment-challenge__content rounded-panel--over-background-banner">
     <ProgressBar @assessment={{@model.assessment}} @answerId={{@model.answer.id}}/>
-    {{component (get-challenge-component-class @model.challenge)
-                challenge=@model.challenge
-                answer=@model.answer
-                assessment=@model.assessment
-                timeoutChallenge=this.timeoutChallenge
-                finishChallenge=this.finishChallenge
-                focusedIn=this.focusedIn
-                focusedOut=this.focusedOut
-                answerValidated=(route-action "saveAnswerAndNavigate")
-                resumeAssessment=(route-action "resumeAssessment")
-                onTooltipClose=this.removeTooltipOverlay
-    }}
+
+    {{#if this.displayTimedChallengeInstructions}}
+      <TimedChallengeInstructions @hasUserConfirmedWarning={{this.setUserConfirmation}} @time={{@model.challenge.timer}} />
+    {{/if}}
+
+    {{#if this.displayChallenge}}
+      {{component (get-challenge-component-class @model.challenge)
+                  challenge=@model.challenge
+                  answer=@model.answer
+                  assessment=@model.assessment
+                  timeoutChallenge=this.timeoutChallenge
+                  finishChallenge=this.finishChallenge
+                  focusedIn=this.focusedIn
+                  focusedOut=this.focusedOut
+                  answerValidated=(route-action "saveAnswerAndNavigate")
+                  resumeAssessment=(route-action "resumeAssessment")
+                  onTooltipClose=this.removeTooltipOverlay
+      }}
+    {{/if}}
   </main>
   <div class="challenge-item__feedback" role="complementary">
     <FeedbackPanel @assessment={{@model.assessment}} @challenge={{@model.challenge}}/>
