+++ mon-pix/app/templates/components/challenge-item-qcu.hbs
@@ -2,60 +2,52 @@
     class="challenge-item"
     data-challenge-id="{{@challenge.id}}">
 
-  {{#if this.isTimedChallengeWithoutAnswer}}
-    {{#unless this.hasUserConfirmedWarning}}
-      <TimedChallengeInstructions @hasUserConfirmedWarning={{this.setUserConfirmation}} @time={{@challenge.timer}} />
-    {{/unless}}
-  {{/if}}
+  <form {{on "submit" this.validateAnswer}} class="rounded-panel rounded-panel--no-margin-bottom">
+    <ChallengeStatement @challenge={{@challenge}} @assessment={{@assessment}} />
 
-  {{#if this.displayChallenge}}
-    <form {{on "submit" this.validateAnswer}} class="rounded-panel rounded-panel--no-margin-bottom">
-      <ChallengeStatement @challenge={{@challenge}} @assessment={{@assessment}} />
-
-      <div class="rounded-panel__row challenge-response {{if @answer 'challenge-response--locked'}} {{if this.hasChallengeTimedOut 'challenge-response--locked'}}">
-        <h2 class="sr-only">{{t 'pages.challenge.parts.answer-input'}}</h2>
-        <div class="challenge-proposals">
-          <QcuProposals @answer={{@answer}}
-                        @answerValue={{@answer.value}}
-                        @proposals={{@challenge.proposals}}
-                        @answerChanged={{this.answerChanged}} />
-        </div>
-
-        {{#if @answer}}
-          <div class="challenge-response__locked-overlay">
-            <FaIcon @icon='lock' class='challenge-response-locked__icon' />
-          </div>
-        {{/if}}
-        {{#if this.hasChallengeTimedOut}}
-          <div class="challenge-response__locked-overlay">
-            <FaIcon @icon='hourglass-end' class='challenge-response-locked__icon' />
-          </div>
-        {{/if}}
-
-        {{#if this.displayTimer}}
-          <div class="timeout-gauge-wrapper">
-            <TimeoutGauge @allottedTime={{@challenge.timer}}
-                          @setChallengeAsTimedOut={{this.setChallengeAsTimedOut}}/>
-          </div>
-        {{/if}}
+    <div class="rounded-panel__row challenge-response {{if @answer 'challenge-response--locked'}} {{if this.hasChallengeTimedOut 'challenge-response--locked'}}">
+      <h2 class="sr-only">{{t 'pages.challenge.parts.answer-input'}}</h2>
+      <div class="challenge-proposals">
+        <QcuProposals @answer={{@answer}}
+                      @answerValue={{@answer.value}}
+                      @proposals={{@challenge.proposals}}
+                      @answerChanged={{this.answerChanged}} />
       </div>
 
-      {{#if this.errorMessage}}
-        <div class="alert alert--danger" role="alert">
-          {{this.errorMessage}}
+      {{#if @answer}}
+        <div class="challenge-response__locked-overlay">
+          <FaIcon @icon='lock' class='challenge-response-locked__icon' />
+        </div>
+      {{/if}}
+      {{#if this.hasChallengeTimedOut}}
+        <div class="challenge-response__locked-overlay">
+          <FaIcon @icon='hourglass-end' class='challenge-response-locked__icon' />
         </div>
       {{/if}}
 
-      {{#if @assessment}}
-        <ChallengeActions @challenge={{@challenge}}
-                          @answer={{@answer}}
-                          @resumeAssessment={{this.resumeAssessment}}
-                          @validateAnswer={{this.validateAnswer}}
-                          @skipChallenge={{this.skipChallenge}}
-                          @isValidateButtonEnabled={{this.isValidateButtonEnabled}}
-                          @hasChallengeTimedOut={{this.hasChallengeTimedOut}}
-                          @isSkipButtonEnabled={{this.isSkipButtonEnabled}}/>
+      {{#if this.displayTimer}}
+        <div class="timeout-gauge-wrapper">
+          <TimeoutGauge @allottedTime={{@challenge.timer}}
+                        @setChallengeAsTimedOut={{this.setChallengeAsTimedOut}}/>
+        </div>
       {{/if}}
-    </form>
-  {{/if}}
+    </div>
+
+    {{#if this.errorMessage}}
+      <div class="alert alert--danger" role="alert">
+        {{this.errorMessage}}
+      </div>
+    {{/if}}
+
+    {{#if @assessment}}
+      <ChallengeActions @challenge={{@challenge}}
+                        @answer={{@answer}}
+                        @resumeAssessment={{this.resumeAssessment}}
+                        @validateAnswer={{this.validateAnswer}}
+                        @skipChallenge={{this.skipChallenge}}
+                        @isValidateButtonEnabled={{this.isValidateButtonEnabled}}
+                        @hasChallengeTimedOut={{this.hasChallengeTimedOut}}
+                        @isSkipButtonEnabled={{this.isSkipButtonEnabled}}/>
+    {{/if}}
+  </form>
 </article>
